requirements.txt
streamlit
pandas
google-cloud-bigquery
google-auth
plotly
db-dtypes
google-cloud-bigquery[all]
